Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> What about a Jenkins job to build this?

CICD is outside the scope of R1 I was told, and **certainly** outside the scope of this PR. The build here is a standard docker build, i'm hoping there is already a JJB for that.

The unit tests won't run without rmr .so files installed, another future point of discussion. 

But again CICD is NOT in the scope of this push.

Patch-set: 1