Update patch set 1
Patch Set 1:

What is the appropriate CI for this?  I see the Dockerfile, suggesting a docker-based build, and I see tox, suggesting a Python-based test.  I do NOT see the Dockerfile invoking tox as part of the build. 

All I'm saying is, what job should be defined that exercises the most code and probes as deeply as possible while building this as a release candidate?

Patch-set: 1