Update patch set 1
Patch Set 1:

Tox is a unit test job. It should NOT run as part of a docker build. That should run as the unit test on every commit job. Every commit (merge request) should ideally run tox; however it requires rmr to be installed so that probably is going to be some work for the LF team.
 
The build job is a normal docker build. This should only be invoked on “build release” or whatever the magical incantation is.

Patch-set: 1