Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> This demonstrates use of a pre-built CI builder image from the o-ran-sc Docker registry.  Don't automatically merge this - let's discuss if this is helpful to you.

For example: there was a critical bug in RMR that affected A1 in Kubernetes. Scott is fixing that today. So I already need to update my rmr commit hash. Is there going to be a base image automatically for every rmr commit hash? If there is, I think this could be helpful as long as I can just point to a new base. Otherwise, this is not helpful.

Patch-set: 1