Update patch set 9
Patch Set 9: Code-Review+1

(1 comment)

> Patch Set 9:
> 
> > Patch Set 9:
> > 
> > (1 comment)
> 
> thanks.
> 
> I'm going to study some more and do another cleanup pass; if you're both comfortable with this being done as a seperate merge request as this is already quite large, I would appreciate a merge here; with a subsequent pull request to follow.

I think you're referring to Lusheng and Dave, but I'll chime in and say I'm good with it :) 

It's also nagging me here about a comment I made on ps2, but apparently I don't know how to add comments as it seems unapplied.  It's not a correction, but an answer to the question about what newrt is.

Patch-set: 9