Update patch set 2
Patch Set 2:

Right, probably have to use -f ci/Dockerfile but that's not gonna find the tag file correctly.  I'm struggling with this little reorg.  I would like to tuck the required CI files away in the ci/ subdir to make a neat & clean organization. However the existing templates don't allow the flexibility that I need.  The templates require starting in a set directory, known as the docker root, but the ci subdir is NOT that.  Stay tuned, this will not be settled today.

Patch-set: 2
Label: Code-Review=0