Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Right, probably have to use -f ci/Dockerfile but that's not gonna find the tag file correctly.  I'm struggling with this little reorg.  I would like to tuck the required CI files away in the ci/ subdir to make a neat & clean organization. However the existing templates don't allow the flexibility that I need.  The templates require starting in a set directory, known as the docker root, but the ci subdir is NOT that.  Stay tuned, this will not be settled today.

I'm still not sure I completely grok the tag purpose, but it sure seems to be a very narrow minded approach to expect the top level to be polluted with the build stuff.  I need to go back and look, but if the tag and jenkins file (my only experience we cobbled together a jenkins file at the top level and that was all we needed) sit at the top level, couldn't the jenkins job to the rest as far as running build with the needed -f (and maybe other) option?

Patch-set: 2
Label: Code-Review=0