Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Right, probably have to use -f ci/Dockerfile but that's not gonna find the tag file correctly.  I'm struggling with this little reorg.  I would like to tuck the required CI files away in the ci/ subdir to make a neat & clean organization. However the existing templates don't allow the flexibility that I need.  The templates require starting in a set directory, known as the docker root, but the ci subdir is NOT that.  Stay tuned, this will not be settled today.

I don't really think moving the dockerfile to a folder makes anything cleaner or neater. Docker projects usually have a dockerfile at the root. To me this just makes this unconventional and I don't really see the benefit.

for the tag, you could make it a .tagfile, but this appears to be an ocd thing for no super benefit

Patch-set: 2