Update patch set 5
Patch Set 5:

the it/dep helm chart currently references 0.8.2. If we merge this as is, this is going to "backdoor" overwrite the docker container they reference. If that is desired, we can merge this as is. If that is not desired, please  bump the version (in all the places dictated by dev-guide)

Patch-set: 5
Label: Code-Review=0