Update patch set 4
Patch Set 4:

I see this error:

17:29:41 /tmp/jenkins5433750432126600069.sh: line 81: docker: command not found

I recall on Friday we changed the A1 jobs including the build node.  I see now that I mistakenly switched ALL jobs away from a node with docker.  That was ok for the sonar job, but not for the docker-image jobs.  Need to change it back to specify a build node that has the docker software installed.

Patch-set: 4