Update patch set 4
Patch Set 4:

as far as I know there is only 1 A1 server in the RIC so I'm not sure what your question means.

I know we are working on the traffic steering use case for ORAN B and Traffic steering will get policy guidance from A1. Are you asking who is going to sit at the keyboard and make the CURL commands of the policy into A1? Idk.

Patch-set: 4