Repair A1 integration tests
Correct COPY command in Dockerfiles to get all files for RMR, not just one.
Pin the version of tavern to 0.34.0, because the latest (1.0.0) changed
the required format of the test-spec file and rejects this one.
Add delay before test AC policy status get to increase chance of success.
Adjust tox ini file for shell-script path and to be a bit more verbose.
Bump version in the a1 chart to 2.1.7.

Signed-off-by: Lott, Christopher (cl778h) <cl778h@att.com>
Change-Id: I96dc018aa80e6bfcbf9c28068129e6cca1611d15