Update patch set 5
Patch Set 5:

This is a step on the way to a happy place where all integration tests pass.  There's no particular rush to mere.  As time permits I will revisit the failing test.  So maybe let this wait a while.

But with that said, I am working on another A1 change set, to address an issue that Ron S pointed out on Friday and to improve documentation, etc.  I will push that soon.

Patch-set: 5