Update patch set 3
Patch Set 3:

This is a gap that Ron S. pointed out.  A1 accepts the policy type ID on the URL as a path parameter, and assumes that the duplicated policy ID in the body is the same.  This change removes the assumption to ensure that the two IDs are the same.  I could have rejected the create on mismatch, but opted for this solution instead.

Patch-set: 3
CC: Gerrit User 1000049 <1000049@84e4e190-e623-4c37-9eda-58b1937c1b53>