Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> This is a gap that Ron S. pointed out.  A1 accepts the policy type ID on the URL as a path parameter, and assumes that the duplicated policy ID in the body is the same.  This change removes the assumption to ensure that the two IDs are the same.  I could have rejected the create on mismatch, but opted for this solution instead.

I agree to the gap but would have rejected this solution.

The real solution would have been to remove this field from the PUT body, since it's in the URL.

THe next best option would have been to throw a 400.

This solution basically takes a guess and just assumes the user got the URL right, when maybe the body is what they wanted and got the URL wrong. You're basically making an assumption about what kind of error the user is likely to have made, which I think is dangerous. Better to remove the possibility for a mismatch, or, if there is a mismatch, throw it back at the user and make them fix it. This "solution" can have unintended effects. -2

Patch-set: 3