Fix conflicting RMR port number
The config.yaml file is configured to use port 4561 as port number for rmrdata that
conflicts with the rmrroute port used to exchange messages with rtmgr.
This issue causes RMR to associate two different file descriptors with port 4561.
As a result, and A1 messages coming from xApps are not processed by A1 mediator.

This change fixes the port number to 4562 in config.yaml file.

Issue-ID: RIC-1012

Change-Id: Ib4231ff1206afce6c1308ed94a7a1b1e265b4cf3
Signed-off-by: Alexandre Huff <alexandrehuff@utfpr.edu.br>